packages/parsers/pyproject.toml
[project]
name = "jurisflow-parsers"
version = "0.1.0"
description = "Document parsing and OCR helpers for Jurisflow."
requires-python = ">=3.11"
dependencies = [
  "numpy>=2.2,<3.0",
  "pillow>=11.0,<12.0",
  "pymupdf>=1.27,<2.0",
  "python-docx>=1.1,<2.0",
]

[project.optional-dependencies]
ocr = [
  "paddleocr>=3.4,<4.0",
  "paddlepaddle>=3.1,<4.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

